Move Enums to data model
