+++ pydeconz/interfaces/lights.py
@@ -10,7 +10,7 @@ from ..models import ResourceGroup, ResourceType
 from ..models.light.configuration_tool import ConfigurationTool
 from ..models.light.cover import Cover
 from ..models.light.fan import Fan
-from ..models.light.light import Light
+from ..models.light.light import Alert, Effect, Light
 from ..models.light.lock import Lock
 from ..models.light.range_extender import RangeExtender
 from ..models.light.siren import Siren
@@ -22,80 +22,12 @@ if TYPE_CHECKING:
 LOGGER = logging.getLogger(__name__)
 
 
-class Alert(enum.Enum):
-    """Temporary alert effect.
-
-    "none" — light is not performing an alert.
-    "lselect" — light is blinking a longer time.
-    "select" — light is blinking a short time.
-    """
-
-    NONE = "none"
-    LONG = "lselect"
-    SHORT = "select"
-
-    # Specific to Hue color bulbs
-
-    BLINK = "blink"
-    BREATHE = "breathe"
-    CHANNEL_CHANGE = "channelchange"
-    FINISH = "finish"
-    OKAY = "okay"
-    STOP = "stop"
-
-    UNKNOWN = "unknown"
-
-    @classmethod
-    def _missing_(cls, value: object) -> "Alert":
-        """Set default enum member if an unknown value is provided."""
-        LOGGER.warning("Unexpected light alert type %s", value)
-        return Alert.UNKNOWN
-
-
 class CoverAction(enum.Enum):
     """Possible cover actions."""
 
     CLOSE = enum.auto()
     OPEN = enum.auto()
     STOP = enum.auto()
-
-
-class Effect(enum.Enum):
-    """Effect of the light.
-
-    "colorloop" — cycle through hue values 0-360.
-    "none" — no effect.
-    """
-
-    COLORLOOP = "colorloop"
-    NONE = "none"
-
-    # Specific to Lidl christmas light
-
-    CARNIVAL = "carnival"
-    COLLIDE = "collide"
-    FADING = "fading"
-    FIREWORKS = "fireworks"
-    FLAG = "flag"
-    GLOW = "glow"
-    RAINBOW = "rainbow"
-    SNAKE = "snake"
-    SNOW = "snow"
-    SPARKLES = "sparkles"
-    STEADY = "steady"
-    STROBE = "strobe"
-    TWINKLE = "twinkle"
-    UPDOWN = "updown"
-    VINTAGE = "vintage"
-    WAVES = "waves"
-
-    UNKNOWN = "unknown"
-
-    @classmethod
-    def _missing_(cls, value: object) -> "Effect":
-        """Set default enum member if an unknown value is provided."""
-        LOGGER.warning("Unexpected light effect type %s", value)
-        return Effect.UNKNOWN
 
 
 class FanSpeed(enum.Enum):
